requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==43.0.1
daphne==4.1.2
distro==1.9.0
dj-rest-auth==6.0.0
Django==5.0.9
django-allauth==0.63.6
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==24.3
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
ecdsa==0.19.0
google-auth==2.32.0
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
jiter==0.5.0
kombu==5.4.1
msgpack==1.1.0
openai==1.45.0
packaging==24.1
pillow==10.4.0
pipdeptree==2.23.1
posuto==2024.7.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
redis==5.0.8
requests==2.32.3
rsa==4.9
schedule==1.2.2
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.66.5
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.0.3
